Add ability to leave a group

+++ convex/conversation.ts
@@ -72,7 +72,7 @@ export const get = query({
   },
 });
 
-export const deleteGroup  = mutation({
+export const deleteGroup = mutation({
   args: {
     conversationId: v.id('conversations'),
   },
@@ -110,14 +110,51 @@ export const deleteGroup  = mutation({
             q => q.eq('conversationId', args.conversationId)).
         collect();
 
-    await ctx.db.delete(args.conversationId)
+    await ctx.db.delete(args.conversationId);
 
     await Promise.all(memberships.map(async membership => {
-      await ctx.db.delete(membership._id)
-    }))
+      await ctx.db.delete(membership._id);
+    }));
 
     await Promise.all(messages.map(async message => {
-      await ctx.db.delete(message._id)
-    }))
+      await ctx.db.delete(message._id);
+    }));
+  },
+});
+
+export const leaveGroup = mutation({
+  args: {
+    conversationId: v.id('conversations'),
+  },
+  handler: async (ctx, args) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error('Unauthorized');
+    }
+
+    const currentUser = await getUserByClerkId({
+      ctx,
+      clerkId: identity.subject,
+    });
+    if (!currentUser) {
+      throw new ConvexError('User not found');
+    }
+
+    const conversation = await ctx.db.get(args.conversationId);
+    if (!conversation) {
+      throw new ConvexError('Conversation not found');
+    }
+
+    const membership = await ctx.db.
+        query('conversationMembers').
+        withIndex('by_memberId_conversationId',
+            q => q.eq('memberId', currentUser._id).
+                eq('conversationId', args.conversationId)).
+        unique();
+    if (!membership) {
+      throw new ConvexError('You are not a member of this conversation');
+    }
+
+    await ctx.db.delete(membership._id);
   },
 });
